requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1 